// src/app/models/escolaridade.ts
export class Escolaridade {
  id: string;
  fundamCompleto: boolean;
  serieFundamental: string;
  medioCompleto: boolean;
  serieMedio: string;
  superior: string;
  cursoSuperior: string;
  pretendeCursar: string;
  horario: string;
  local: string;
  curriculoId: number;
  curriculo: number;
}